moves sendMessage to util/messenger
